+++ apps/prairielearn/elements/pl-order-blocks/dag_checker.py
@@ -1,6 +1,6 @@
 import itertools
 from collections import Counter
-from collections.abc import Iterable, Mapping
+from collections.abc import Iterable, Mapping, Sequence
 from copy import deepcopy
 
 import networkx as nx
@@ -31,7 +31,8 @@ def validate_grouping(
 def solve_dag(
     depends_graph: Mapping[str, list[str]], group_belonging: Mapping[str, str | None]
 ) -> list[str]:
-    """Solve the given problem
+    """
+    Solve the given problem
     :param depends_graph: The dependency graph between blocks specified in the question
     :param group_belonging: which pl-block-group each block belongs to, specified in the question
     :return: a list that is a topological sort of the input DAG with blocks in the same group occuring
@@ -54,7 +55,9 @@ def solve_dag(
     return sort
 
 
-def check_topological_sorting(submission: list[str], graph: nx.DiGraph) -> int:
+def check_topological_sorting(
+    submission: Sequence[str | None], graph: nx.DiGraph
+) -> int:
     """
     :param submission: candidate for topological sorting
     :param graph: graph to check topological sorting over
@@ -69,7 +72,7 @@ def check_topological_sorting(submission: list[str], graph: nx.DiGraph) -> int:
 
 
 def check_grouping(
-    submission: list[str], group_belonging: Mapping[str, str | None]
+    submission: Sequence[str | None], group_belonging: Mapping[str, str | None]
 ) -> int:
     """
     :param submission: candidate solution
@@ -81,7 +84,7 @@ def check_grouping(
     cur_group = None
     cur_group_size = 0
     for i, node in enumerate(submission):
-        group_id = group_belonging.get(node)
+        group_id = None if node is None else group_belonging.get(node)
         if group_id is not None and cur_group is None:
             cur_group = group_id
         elif group_id is None and cur_group is not None:
@@ -149,11 +152,12 @@ def add_edges_for_groups(
 
 
 def grade_dag(
-    submission: list[str],
+    submission: Sequence[str | None],
     depends_graph: Mapping[str, list[str]],
     group_belonging: Mapping[str, str | None],
 ) -> tuple[int, int]:
-    """In order for a student submission to a DAG graded question to be deemed correct, the student
+    """
+    In order for a student submission to a DAG graded question to be deemed correct, the student
     submission must be a topological sort of the DAG and blocks which are in the same pl-block-group
     as one another must all appear contiguously.
     :param submission: the block ordering given by the student
@@ -171,7 +175,8 @@ def grade_dag(
 
 
 def is_vertex_cover(G: nx.DiGraph, vertex_cover: Iterable[str]) -> bool:
-    """this function from
+    """
+    Taken from
     https://docs.ocean.dwavesys.com/en/stable/docs_dnx/reference/algorithms/generated/dwave_networkx.algorithms.cover.is_vertex_cover.html
     """
     cover = set(vertex_cover)
@@ -179,11 +184,12 @@ def is_vertex_cover(G: nx.DiGraph, vertex_cover: Iterable[str]) -> bool:
 
 
 def lcs_partial_credit(
-    submission: list[str],
+    submission: Sequence[str | None],
     depends_graph: Mapping[str, list[str]],
     group_belonging: Mapping[str, str | None],
 ) -> int:
-    """Computes the number of edits required to change the student solution into a correct solution using
+    """
+    Compute the number of edits required to change the student solution into a correct solution using
     largest common subsequence edit distance (allows only additions and deletions, not replacing).
     The naive solution would be to enumerate all topological sorts, then get the edit distance to each of them,
     but this would be too slow. Instead, our algorithm is as follows:
@@ -204,7 +210,9 @@ def lcs_partial_credit(
     """
     graph = dag_to_nx(depends_graph, group_belonging)
     trans_clos = nx.transitive_closure(graph)
-    submission_no_distractors = [node for node in submission if node in depends_graph]
+    submission_no_distractors = [
+        node for node in submission if node in depends_graph and node is not None
+    ]
 
     # if node1 must occur before node2 in any correct solution, but node2 occurs before
     # node1 in the submission, add them both and an edge between them to the problematic subgraph