+++ apps/prairielearn/elements/pl-rich-text-editor/pl-rich-text-editor.py
@@ -25,6 +25,7 @@ ALLOW_BLANK_DEFAULT = False
 SOURCE_FILE_NAME_DEFAULT = None
 DIRECTORY_DEFAULT = "."
 MARKDOWN_SHORTCUTS_DEFAULT = True
+CLIPBOARD_ENABLED_DEFAULT = True
 
 
 def get_answer_name(file_name: str) -> str:
@@ -34,9 +35,9 @@ def get_answer_name(file_name: str) -> str:
 
 
 def element_inner_html(element: lxml.html.HtmlElement) -> str:
-    return (element.text or "") + "".join(
-        [str(lxml.html.tostring(c), "utf-8") for c in element.iterchildren()]
-    )
+    return (element.text or "") + "".join([
+        str(lxml.html.tostring(c), "utf-8") for c in element.iterchildren()
+    ])
 
 
 def prepare(element_html: str, data: pl.QuestionData) -> None:
@@ -51,6 +52,7 @@ def prepare(element_html: str, data: pl.QuestionData) -> None:
         "format",
         "markdown-shortcuts",
         "counter",
+        "clipboard-enabled",
     ]
     pl.check_attribs(element, required_attribs, optional_attribs)
     source_file_name = pl.get_string_attrib(
@@ -95,6 +97,9 @@ def render(element_html: str, data: pl.QuestionData) -> str:
         element, "markdown-shortcuts", MARKDOWN_SHORTCUTS_DEFAULT
     )
     counter = pl.get_enum_attrib(element, "counter", Counter, Counter.NONE)
+    clipboard_enabled = pl.get_boolean_attrib(
+        element, "clipboard-enabled", CLIPBOARD_ENABLED_DEFAULT
+    )
     element_text = element_inner_html(element)
 
     if data["panel"] == "question" or data["panel"] == "submission":
@@ -115,6 +120,7 @@ def render(element_html: str, data: pl.QuestionData) -> str:
             "markdown_shortcuts": "true" if markdown_shortcuts else "false",
             "counter": counter.value,
             "counter_enabled": counter != Counter.NONE,
+            "clipboard_enabled": clipboard_enabled,
         }
 
         if source_file_name is not None:
@@ -127,10 +133,8 @@ def render(element_html: str, data: pl.QuestionData) -> str:
             file_path = os.path.join(directory, source_file_name)
             with open(file_path) as f:
                 text_display = f.read()
-        elif element_text is not None:
-            text_display = str(element_text)
         else:
-            text_display = ""
+            text_display = element_text
 
         html_params["original_file_contents"] = base64.b64encode(
             text_display.encode("UTF-8").strip()