+++ apps/prairielearn/python/prairielearn/internal/traverse.py
@@ -1,6 +1,7 @@
 from collections import deque
 from collections.abc import Callable
 from html import escape as html_escape
+from html import unescape as html_unescape
 from itertools import chain
 
 import lxml.html
@@ -8,23 +9,21 @@ import lxml.html
 ElementReplacement = str | lxml.html.HtmlElement | list[lxml.html.HtmlElement] | None
 
 # https://developer.mozilla.org/en-US/docs/Glossary/Void_element
-VOID_ELEMENTS = frozenset(
-    {
-        "area",
-        "base",
-        "br",
-        "col",
-        "embed",
-        "hr",
-        "img",
-        "input",
-        "link",
-        "meta",
-        "source",
-        "track",
-        "wbr",
-    }
-)
+VOID_ELEMENTS = frozenset({
+    "area",
+    "base",
+    "br",
+    "col",
+    "embed",
+    "hr",
+    "img",
+    "input",
+    "link",
+    "meta",
+    "source",
+    "track",
+    "wbr",
+})
 
 UNESCAPED_ELEMENTS = frozenset({"script", "style"})
 
@@ -69,7 +68,6 @@ def traverse_and_replace(
     The top entry in count_stack is decremented every time something is moved onto result,
     and when an entry hits zero, the corresponding tag from tail_stack is moved onto result as well.
     """
-
     # Initialize result and work data structures
     result: deque[str] = deque()
 
@@ -93,7 +91,7 @@ def traverse_and_replace(
                 new_elements = []
             elif isinstance(new_elements, str):
                 fragments = lxml.html.fragments_fromstring(new_elements)
-                new_elements = fragments if fragments is not None else []
+                new_elements = fragments
 
             if isinstance(new_elements, list):
                 # Modify count stack for new elements and decrement for element that was replaced
@@ -138,7 +136,17 @@ def traverse_and_replace(
                     if new_elements.tag in UNESCAPED_ELEMENTS:
                         result.append(new_elements.text)
                     else:
-                        result.append(html_escape(new_elements.text))
+                        # `lxml` uses `libxml2` under the hood, which does not support
+                        # the full set of HTML5 named entities:
+                        # https://gitlab.gnome.org/GNOME/libxml2/-/issues/857
+                        # This means that with a naive approach, we'd end up double
+                        # escaping entities like `&langle;` into `&amp;langle;`. To work
+                        # around this (at least until `libxml2` hopefully adds support for
+                        # HTML5 named entities), we first unescape the text to get the
+                        # actual Unicode characters, and then escape them again. Escaping
+                        # will only escape `&`, `<`, and `>`; it won't escape everything
+                        # that could possibly be represented by a named entity.
+                        result.append(html_escape(html_unescape(new_elements.text)))
 
                 # Add all children to the work stack
                 children = list(new_elements)
@@ -154,7 +162,7 @@ def traverse_and_replace(
             count_stack.pop()
             tail_tag, tail_text = tail_stack.pop()
 
-            if tail_tag not in VOID_ELEMENTS and tail_tag is not None:
+            if tail_tag not in VOID_ELEMENTS:
                 result.append(f"</{tail_tag}>")
             if tail_text is not None:
                 result.append(tail_text)