+++ apps/prairielearn/python/test/core_test.py
@@ -12,6 +12,7 @@ import numpy as np
 import pandas as pd
 import prairielearn as pl
 import pytest
+from numpy.typing import ArrayLike
 
 
 def city_dataframe() -> pd.DataFrame:
@@ -78,27 +79,25 @@ def breast_cancer_dataframe() -> pd.DataFrame:
 
 
 def r_types_dataframe() -> pd.DataFrame:
-    return pd.DataFrame(
-        {
-            # Scalars
-            "integer": 1,
-            "numeric": 3.14,
-            "logical": False,
-            "character": "foo",
-            # TODO adding in complex numbers won't deserialize correctly, fix this (somehow?)
-            # "complex": complex(1, 2),
-            # Series
-            "numeric-list": pd.Series([1.0] * 3, dtype="float64"),
-            "integer-list": pd.Series([1] * 3, dtype="int64"),
-            # "complex-list": pd.Series(np.array([1, 2, 3]) + np.array([4, 5, 6]) *1j).astype("complex128"),
-            "character-list": pd.Series(["hello", "world", "stat"]),
-            "logical-list": pd.Series([True, False, True]),
-            "character-string-list": pd.Series(["a", "b", "c"]),
-            # Time Dependency: https://pandas.pydata.org/docs/user_guide/timeseries.html
-            "POSIXct-POSIXt-timestamp": pd.Timestamp("2023-01-02T00:00:00.0000000"),
-            "POSIXct-POSIXt-date_range": pd.date_range("2023", freq="D", periods=3),
-        }
-    )
+    return pd.DataFrame({
+        # Scalars
+        "integer": 1,
+        "numeric": 3.14,
+        "logical": False,
+        "character": "foo",
+        # TODO adding in complex numbers won't deserialize correctly, fix this (somehow?)
+        # "complex": complex(1, 2),
+        # Series
+        "numeric-list": pd.Series([1.0] * 3, dtype="float64"),
+        "integer-list": pd.Series([1] * 3, dtype="int64"),
+        # "complex-list": pd.Series(np.array([1, 2, 3]) + np.array([4, 5, 6]) *1j).astype("complex128"),
+        "character-list": pd.Series(["hello", "world", "stat"]),
+        "logical-list": pd.Series([True, False, True]),
+        "character-string-list": pd.Series(["a", "b", "c"]),
+        # Time Dependency: https://pandas.pydata.org/docs/user_guide/timeseries.html
+        "POSIXct-POSIXt-timestamp": pd.Timestamp("2023-01-02T00:00:00.0000000"),
+        "POSIXct-POSIXt-date_range": pd.date_range("2023", freq="D", periods=3),
+    })
 
 
 @pytest.mark.parametrize(
@@ -145,7 +144,9 @@ def test_encoding_legacy(df: pd.DataFrame) -> None:
         nx.gnc_graph(20),
     ],
 )
-def test_networkx_serialization(networkx_graph: Any) -> None:
+def test_networkx_serialization(
+    networkx_graph: nx.Graph | nx.DiGraph | nx.MultiGraph | nx.MultiDiGraph,
+) -> None:
     """Test equality after conversion of various numpy objects."""
 
     networkx_graph.graph["rankdir"] = "TB"
@@ -160,7 +161,10 @@ def test_networkx_serialization(networkx_graph: Any) -> None:
     json_object = json.dumps(pl.to_json(networkx_graph), allow_nan=False)
     decoded_json_object = pl.from_json(json.loads(json_object))
 
-    assert type(networkx_graph) == type(decoded_json_object)  # noqa: E721
+    assert type(decoded_json_object) is type(networkx_graph)
+
+    # This check is needed because pyright cannot infer the type of decoded_json_object
+    assert isinstance(decoded_json_object, type(networkx_graph))
 
     assert nx.utils.nodes_equal(networkx_graph.nodes(), decoded_json_object.nodes())
     assert nx.utils.edges_equal(networkx_graph.edges(), decoded_json_object.edges())
@@ -238,7 +242,7 @@ def test_set_score_data(
         np.ones((2, 3, 4), dtype=np.int16),
     ],
 )
-def test_numpy_serialization(numpy_object: Any) -> None:
+def test_numpy_serialization(numpy_object: ArrayLike) -> None:
     """Test equality after conversion of various numpy objects."""
 
     json_object = json.dumps(
@@ -246,7 +250,11 @@ def test_numpy_serialization(numpy_object: Any) -> None:
     )
     decoded_json_object = pl.from_json(json.loads(json_object))
 
-    assert type(numpy_object) == type(decoded_json_object)  # noqa: E721
+    assert type(numpy_object) is type(decoded_json_object)
+
+    # This check is needed because pyright cannot infer the type of decoded_json_object
+    assert isinstance(decoded_json_object, type(numpy_object))
+
     np.testing.assert_array_equal(numpy_object, decoded_json_object, strict=True)
 
 
@@ -435,3 +443,104 @@ def test_iter_keys(length: int, expected_output: list[str]) -> None:
 )
 def test_index2key(idx: int, expected_output: str) -> None:
     assert pl.index2key(idx) == expected_output
+
+
+@pytest.mark.parametrize(
+    ("value", "args", "expected_output"),
+    [
+        (0, {}, "0.00"),
+        (0, {"digits": 1}, "0.0"),
+        (0, {"digits": 0}, "0"),
+        (0.0, {}, "0.00"),
+        (0.0, {"digits": 1}, "0.0"),
+        (0.0, {"digits": 0}, "0"),
+        (np.float64(0.0), {}, "0.00"),
+        (np.float64(0.0), {"digits": 1}, "0.0"),
+        (np.float64(0.0), {"presentation_type": "sigfig"}, "0.0"),
+        (np.zeros(2), {}, "[0.00, 0.00]"),
+        (np.zeros(2), {"digits": 1}, "[0.0, 0.0]"),
+        (np.zeros(2), {"digits": 0}, "[0, 0]"),
+        (np.zeros(2), {"language": "matlab"}, "[0.00, 0.00]"),
+        (np.zeros(2), {"language": "mathematica"}, "{0.00, 0.00}"),
+        (np.zeros(2), {"language": "r"}, "c(0.00, 0.00)"),
+        (np.zeros(2), {"language": "sympy"}, "Matrix([0.00, 0.00])"),
+        (np.zeros((2, 2)), {}, "[[0.00, 0.00], [0.00, 0.00]]"),
+        (np.zeros((2, 2)), {"digits": 1}, "[[0.0, 0.0], [0.0, 0.0]]"),
+        (np.zeros((2, 2)), {"digits": 0}, "[[0, 0], [0, 0]]"),
+        (np.zeros((2, 2)), {"language": "matlab"}, "[0.00 0.00; 0.00 0.00]"),
+        (np.zeros((2, 2)), {"language": "mathematica"}, "{{0.00, 0.00}, {0.00, 0.00}}"),
+        (
+            np.zeros((2, 2)),
+            {"language": "r"},
+            "matrix(c(0.00, 0.00, 0.00, 0.00), nrow = 2, ncol = 2, byrow = TRUE)",
+        ),
+        (
+            np.zeros((2, 2)),
+            {"language": "sympy"},
+            "Matrix([[0.00, 0.00], [0.00, 0.00]])",
+        ),
+    ],
+)
+def test_string_from_numpy(value: Any, args: dict, expected_output: str) -> None:
+    assert pl.string_from_numpy(value, **args) == expected_output
+
+
+@pytest.mark.parametrize(
+    ("value", "args", "expected_output"),
+    [
+        (0, {}, "0.0"),
+        (0, {"digits": 1}, "0."),
+        (0, {"digits": 0}, "0"),
+        (0.0, {}, "0.0"),
+        (0.0, {"digits": 1}, "0."),
+        (0.0, {"digits": 0}, "0"),
+        (complex(1, 2), {}, "1.0+2.0j"),
+        (complex(0, 2), {}, "0.0+2.0j"),
+        (complex(1, 0), {}, "1.0+0.0j"),
+        (np.complex64(complex(1, 2)), {}, "1.0+2.0j"),
+        (np.complex64(complex(0, 2)), {}, "0.0+2.0j"),
+        (np.complex64(complex(1, 0)), {}, "1.0+0.0j"),
+        # For legacy reasons, we must also support strings.
+        ("0", {}, "0.0"),
+        ("0", {"digits": 1}, "0."),
+        ("0", {"digits": 0}, "0"),
+    ],
+)
+def test_string_from_number_sigfig(
+    value: Any, args: dict, expected_output: str
+) -> None:
+    assert pl.string_from_number_sigfig(value, **args) == expected_output
+
+
+@pytest.mark.parametrize(
+    ("value", "args", "expected_output"),
+    [
+        (0, {}, "0.0"),
+        (0, {"ndigits": 1}, "0."),
+        (0, {"ndigits": 0}, "0"),
+        (0.0, {}, "0.0"),
+        (0.0, {"ndigits": 1}, "0."),
+        (0.0, {"ndigits": 0}, "0"),
+        (complex(1, 2), {}, "1.0+2.0j"),
+        (complex(0, 2), {}, "0.0+2.0j"),
+        (complex(1, 0), {}, "1.0+0.0j"),
+        (np.complex64(complex(1, 2)), {}, "1.0+2.0j"),
+        (np.complex64(complex(0, 2)), {}, "0.0+2.0j"),
+        (np.complex64(complex(1, 0)), {}, "1.0+0.0j"),
+    ],
+)
+def test_numpy_to_matlab_sf(value: Any, args: dict, expected_output: str) -> None:
+    assert pl.numpy_to_matlab_sf(value, **args) == expected_output
+
+
+@pytest.mark.parametrize(
+    ("value", "expected_output"),
+    [
+        (0, "0.00"),
+        (0.0, "0.00"),
+        (complex(1, 2), "1.00+2.00j"),
+        (np.array([[1, 2], [3, 4]]), r"\begin{bmatrix}  1 & 2\\  3 & 4\\\end{bmatrix}"),
+    ],
+)
+def test_latex_from_2darray(value: Any, expected_output: str) -> None:
+    assert pl.latex_from_2darray(value) == expected_output